requirements.txt
requests
pillow
django-crispy-forms
crispy-bootstrap5